backend part completly added

+++ fullstack-crud/backend/src/controllers/note.controller.js
@@ -25,24 +25,33 @@ router.post('/create', async(req, res) => {
 
 router.patch('/edit/:id', async(req, res) => {
     //error here
-    const note1 = await Notes.findById(req.params.id)
-    const userID_in_note=note1.userID;
+    const id = req.params.id;
+    const note = await Notes.findOne({"_id": id});
+    const userID_in_note=note.userID;
     const userID_making_req=req.body.userID;
     try{
         if(userID_making_req !== userID_in_note){
-            res.status(201).send({"msg":"you are not authoriazed"})
+            res.status(500).send({"msg":"you are not authoriazed"})
         }else{
-            const note=await Notes.findByIdAndUpdate(req.params.id, req.body, {new:true}).lean().exec();
-            res.status(200).send(note);
+            const notes = await Notes.findByIdAndUpdate(req.params.id, req.body, {new:true}).lean().exec();
+            res.status(200).send(notes);
         }
     }catch(err){
         res.status(500).send({message:err.message})
     }
 })
 router.delete('/delete/:id', async(req, res) => {
+    const id = req.params.id;
+    const note = await Notes.findOne({"_id": id});
+    const userID_in_note=note.userID;
+    const userID_making_req=req.body.userID;
     try{
-        const note=await Notes.findByIdAndDelete(req.params.id).lean().exec();
-        res.status(200).send(note);
+        if(userID_making_req !== userID_in_note){
+            res.status(500).send({"msg":"you are not authoriazed"})
+        }else{
+            const notes = await Notes.findByIdAndDelete(req.params.id).lean().exec();
+            res.status(200).send(notes);
+        }
     }catch(err){
         res.status(500).send({message:err.message})
     }
